get_inputs

+++ im2txt/ops/rnn_cell_ops.py
@@ -343,7 +343,7 @@ class BasicLSTMCell(RNNCell):
       z_i = array_ops.zeros([batch_size,subfeature_length])
       state_length = self._num_units
       # with vs.variable_scope(scope or type(self).__name__,initializer=self._initializer):
-      f_att_matrix = vs.get_variable(name="f_att_matrix",shape = (subfeature_length,state_length), initializer=tf.contrib.layers.xavier_initializer())
+      f_att_matrix = vs.get_variable(name="f_att_matrix",shape = (subfeature_length,state_length), initializer=tf.contrib.layers.xavier_initializer(),dtype=tf.float32)
         
       word_imbeddings=inputs[:,0:word_imbedding_length]
       if single_input_length != word_imbedding_length:
